feature: add field creation and auto-focus on enter key in tables
