+++ src/components/EditorSidePanel/TablesTab/TableInfo.jsx
@@ -15,6 +15,7 @@ import TableField from "./TableField";
 import IndexDetails from "./IndexDetails";
 import { useTranslation } from "react-i18next";
 import { dbToTypes } from "../../../data/datatypes";
+import { createNewField } from "./createNewField";
 
 import { useNavigate } from "react-router-dom";
 
@@ -325,104 +326,22 @@ export default function TableInfo({ data }) {
           </Button>
           <Button
             onClick={() => {
-              setUndoStack((prev) => [
-                ...prev,
-                {
-                  action: Action.EDIT,
-                  element: ObjectType.TABLE,
-                  component: "field_add",
-                  tid: data.id,
-                  message: t("edit_table", {
-                    tableName: data.name,
-                    extra: "[add field]",
-                  }),
-                },
-              ]);
-              setRedoStack([]);
-
-              const incr = data.increment && !!dbToTypes[database][settings.defaultFieldType].canIncrement;
-              // Function to get the default size configured by the user
-              const getUserDefaultSize = (typeName) => {
-                const dbSettings = settings?.defaultTypeSizes?.[database] || {};
-                const userSize = dbSettings[typeName];
-                if (typeof userSize === 'number') {
-                  return userSize;
-                }
-                return dbToTypes[database][typeName]?.defaultSize || '';
-              };
-              // Function to get the combined size for types with precision and scale
-              const getUserDefaultPrecisionScale = (typeName) => {
-                const dbSettings = settings?.defaultTypeSizes?.[database] || {};
-                const userSettings = dbSettings[typeName];
-                if (typeof userSettings === 'object') {
-                  const precision = userSettings?.precision || 10;
-                  const scale = userSettings?.scale;
-                  // If it has a defined scale, combine as "precision,scale"
-                  if (scale !== undefined && scale !== null) {
-                    return `${precision},${scale}`;
-                  }
-                  // If it only has precision, return just the precision
-                  return precision.toString();
-                }
-                // Default value for types with precision
-                return "10";
-              };
-
-              // Base field data
-              const newFieldData = {
-                name: "",
-                type: settings.defaultFieldType,
-                default: "",
-                check: "",
-                primary: false,
-                unique: false,
-                notNull: settings.defaultNotNull,
-                increment: false,
-                comment: "",
-                foreignK: false,
-              };
-
-              // Field updates based on type
-              let fieldUpdates = {
-                increment: incr,
-              };
-
-              if (settings.defaultFieldType === "ENUM" || settings.defaultFieldType === "SET") {
-                fieldUpdates = {
-                  ...fieldUpdates,
-                  values: data.values ? [...data.values] : [],
-                };
-              } else if (dbToTypes[database][settings.defaultFieldType].hasPrecision) {
-                fieldUpdates = {
-                  ...fieldUpdates,
-                  size: getUserDefaultPrecisionScale(settings.defaultFieldType),
-                };
-              } else if (dbToTypes[database][settings.defaultFieldType].isSized) {
-                fieldUpdates = {
-                  ...fieldUpdates,
-                  size: getUserDefaultSize(settings.defaultFieldType),
-                };
-              } else if (!dbToTypes[database][settings.defaultFieldType].hasDefault || incr) {
-                fieldUpdates = {
-                  ...fieldUpdates,
-                  default: "",
-                  size: "",
-                  values: [],
-                };
-              } else if (dbToTypes[database][settings.defaultFieldType].hasCheck) {
-                fieldUpdates = {
-                  ...fieldUpdates,
-                  check: "",
-                };
-              } else {
-                fieldUpdates = {
-                  ...fieldUpdates,
-                  size: "",
-                  values: [],
-                };
-              }
-              // Use the new atomic function
-              addFieldToTable(data.id, newFieldData, fieldUpdates);
+                createNewField({
+                    data,
+                    settings,
+                    database,
+                    dbToTypes,
+                    addFieldToTable,
+                    setUndoStack,
+                    setRedoStack,
+                    t,
+                    tid: data.id,
+                });
+                // Focus the new field input, using setTimeout to ensure its rendered
+                setTimeout(() => {
+                    const input = document.getElementById(`scroll_table_${data.id}_input_${1}`);
+                    if (input) input.focus();
+                }, 0);
             }}
             block
           >
